requirements.txt
openai>=1.50.0
python-dotenv>=1.0.1
rich>=13.7.1
playwright>=1.44.0
